Enable snapshot importing (#47)
* Init snapshot importing cmd

* Fix account parser

* Rename config account

* Fix clienttype into clienttypes

* Fix name

* Fix name

* Implement snapshot cmd

* Fix uint64 in pq database bug

* Update token balances

* Remove unused account in snapshot importing

* Apply goroutine pool in snapshot

+++ db/schema/02-token.sql
@@ -10,9 +10,9 @@ CREATE INDEX token_authority_index ON token (mint_authority);
 
 CREATE TABLE token_supply
 (
-    mint    TEXT    NOT NULL PRIMARY KEY,
-    slot    BIGINT  NOT NULL,
-    supply  BIGINT  NOT NULL
+    mint    TEXT            NOT NULL PRIMARY KEY,
+    slot    BIGINT          NOT NULL,
+    supply  NUMERIC(20,0)   NOT NULL
 );
 
 CREATE TABLE token_account
